Use text-area for description input
Currently when we are add/editing text for our function descriptions, the text box uses Codex's text-input which restricts the user from seeing what they've entered as the box does not allow for expansion when it exceeds a certain number of chars. We want to use text-area for better usability for such cases. Additonally, we want the 'char count' to be dynamically located dependent on the language; one that reads LTR vs RTL.

Bug: T353236
Change-Id: I86ce4a24b28d40a39ad8504230f3e981f7ddcedf