Add fetching state as a local flag instead of global getter
Since adding parsers/renderers, the test results API is run also for
generating the renderer field examples and placeholders. This means that
this api can be called from different places in the same page with also
different values. The flag "fetching" is only necessary for the
component FunctionReport, so this patch converts it into a local flag.

Bug: T360018
Change-Id: If183b6644a9077566fa26e807e4a313330d8ad2e