Correct the edit link of an old revision
Right now when we are viewing an [old] revision of a Zobject, the 'edit' button links us to the wrong location in that it does not take us to the page of the respective 'oldid'. In order be routed correctly, we need to make sure that 'oldid' is included in said 'edit' query.

Bug: T343653
Change-Id: Ic72a4a8eed8c477f40f0f467eba8368ce968b000