ZObjectContentHandler::createZObjectViewHeader: Use target's label for chip
We were using the wrong target object label – we were deciding to show a
chip for the label based on whether its Type had a label in the user's
view language, which mean that e.g. if a Function had a label in Latin
but the Function Type itself hadn't been translated into Latin, we'd put
a 'en' chip there saying we think the label was in English, when actually
it was in Latin.

We also were showing the 'Chip' widget twice on the main Title but then
blanking it when duplicate, rather than associating one Chip with each
label, either in the title for the label or the subtitle for the type;
now we show them in the right place. This tweaks the styling so that the
Chip isn't italicised or made grey, but is still small and in-line with
the subtitle.

Additional changes:
 - Rename all the variables in this method to make it clearer which is
   which, which should avoid this outcome recurring.
 - Rename all 'iso' concepts to BCP47, which they nominally are
   This is the more normal label we use at Wikimedia for this specific
   standard, even if sometimes they're actually non-standard MW language
   codes. This also covers parallel code changes in ZObjectEditAction.
 - Drop the unnecessary getIsoCodeIfUserLangIsDifferent() method
   All callers were, or should have been, already deciding whether or
   not to call this anyway.
 - Rename testGetIsoCode() to testWrapBCP47CodeInFakeCodexChip()
   This doesn't get a code, it gets a Codex-style 'Chip' widget (but not
   yet from Codex, hence the name).
 - Added an explicit test that the language Chip for the Type shows up
   when inheriting, using pcd/Picard, and tweak existing tests to check
   that the expected number and nature of Chip widgets are shown.

Bug: T360000
Change-Id: Ife89133ffd7088da145a34680b5659bb605879ee