Make edit page title responsive to edit updates
When we are editing an implementation, the header/title does not update according to any changes made during the edit (this is not an issue when we are editing a Function), particularly when we open the 'About' dialog. As a user, I would want to see the updates I've made as I am editing not only until I have 'published' it. Additionally, if the title happens to be blank, aka. 'Untitled', we want to make sure that any valid update to the title is a standard bold font and not remain gray'd out as if still 'Untitled'.

Bug: T358209
Change-Id: Ia228c85844b83dea6305bcd250fc5169f5424344