Use text-area for description for About widget
Currently when we are editing text for our function descriptions using the About widget dialog, the text box uses Codex's text-input which restricts the user from seeing what they've entered as the box does not allow for expansion when it exceeds a certain number of chars. We want to use text-area for better usability for such cases. Additonally, we want the 'char count' to be dynamically located dependent on the language; one that reads LTR vs RTL.

Bug: T359158
Change-Id: I3d5694d52f75a3728ed9c70d165012a5d25ebc16