Action APIs: Set most of our APIs to emit a cache header for 24 hours
Note that MediaWiki out-guesses us and over-rides the 'public' value
back to 'anon-public-user-private', which is at least better than it
was as 'private'.

Not setting a cache for the perform-test API end-point, as those are
not expected to be run often anyway (as the client is inside the PHP
code for WikiLambda, and cached into the database there).

Setting the public function call result cache value to 7 days, not 1.

Bug: T362271
Change-Id: Ife1893822a44f2523d376ac9eef773fa3191bd8f