selenium: Fixing broken and flaky tests in connect.js
Adding conditionals to connect and disconnect tests and implementations

Waiting dynamically for a class to be removed before rechecking the
checkbox and then clicking the connect button.  The timing of the
checkbox unchecking is different in the UI from when the status changes
from "connected" to "disconnected" and vice versa in the table row UI
so special wait conditions needed to be added.

Bug: T349836
Change-Id: I771e426f9fb37cbf51664d5514f266d3a6c5926e