SkipLink + CreateObject: Fix jumping to content deleting labels in Chrome
Clicking on the MW SkipLink adds a hash #bodyContent to the url and places the focus in the vue app.
Hashchanges in Chrome (FF does not) fire popstate.
Because of this, the initializeView is called again and again, adding top rows and zobjects to the state.
Each SkipLink click doubles the zobject state.

We can not remove the popstate because it is still used:
- when a Function is chose as a Type on the CreateObject page; ZObjectKeyValue routes to the function-editor using history.pushState
- When pressing the back button to go back to the CreateObject page, the popstate is fired.
- The popstate needs to re-initialize the CreateObject page with a clean object, so the popstate is usefull here

Solution:
1. Always make sure the zobject state is cleared when the initializeView is called.
2. Do not call initializeView when there is a hash in the url
3. Remove the hash after poppstate is fired, so it doesnt persist and will not be called twice preventing it from navigating back to CreateObject from the function-editor

After fixing the popstate issue I also noticed leaveTo of Publish.vue got called when clicking the skiplink, because it thinks its navigating away from the current page.
Solution: Check if the href is the same as the current page, if so, do not call leaveTo.

Bug: T348094
Change-Id: I679ec0659d36d29e2df48fe7d40af7ed835efb4a