Add findFunctionsByIOTypes for use with DB table wikilambda_zobject_join
This method provides a comprehensive way to find Wikifunctions
functions having inputs and output of the given types.   Because it
constructs a single query (as opposed to a separate query for each
given I/O type), this approach can take advantage of relevant
optimizations provided by the DBMS. It benefits from the declared
index wlzo_index_key_related_zobject.

Bug: T301712
Change-Id: Ie55b2ad0e7d85bfe265175a61fa6b90eecbb9d77