selenium: Change selectors to account for Codex component changes
When a new codex component was added in, this affected the Selenium
tests.  This patch fixes a couple of the issues, and for InputDropdown.js removes unnecessary specificity that relied on codex code that could change upstream.  As noted in the comment in ContentBlock.js, .cdx-menu__listbox is from an upstream codex component and is not in the WikiLambda codebase.

Bug: T372415
Change-Id: I48e2735a3190d8398c61dec13e753c8ce034629f
Depends-On: I107a6dcb7242d2ae7a29b22c19ae0f617dc9a188