FunctionEditor: Do not fire change event when component will unmount
This is a common issue when dealing with forms or input fields in a web application,
especially when navigating away from a page or a component with unsaved input.
In this case, when the popstate event is triggered, the input field is removed from the DOM,
causing the change event to fire with the value of the input field that is currently focussed.
This causes inconsistent behaviour because on change values are persisted.

Bug: T369713
Change-Id: I7e586cc4e6af2a53c29c2d65bcf8c62b6ed2a805