Initial VisualEditor set-up
This creates the editor with a dialog with two entries, 'function'
and 'inputs', both plain text inputs, with the latter conjoined with
'|' characters for multiple values (e.g. 'Z802' and 'Z41|foo|bar').

We need to replace the dialog contents with a proper Vue WL surface,
or at least build something better, but this gets the initial node
to a state that should be marginally editable, albeit only by experts.

The new i18n messages are split out into their own folder, for easier
prioritisation by translators, though most of the messages shown to
users will in the end be from the main application i18n block.

NOTE: This code is written presuming the wikis it's used on have the
new Parsoid DOM content structure (T373253) enabled for our parser
function; the HTML read and generated by the VE code in this patch
will not be compatible with any wikis using the default mode, until
Parsoid provides a mechanism by which we can individually register
our parser function as using the new DOM syntax and not the old one.

Bug: T373118
Change-Id: I64c42b2467542a4d29a774e2792bf1ef84460d99