About.vue: the "Cancel" button in "Publish your changes" launched from the About dialog loses all the changes in view mode
When pressing edit (pencil) in view mode of About Widget, the user is taken to the edit mode of about widet.
If the user then presses "Publish" and then "Cancel" in the "Publish your changes" dialog, all the changes are lost.
This is not what is happening in edit mode. There the fields stay open after cancelling publish.

I think according to this ticket users find it unexpected that their changes are lost.
(Although this is about the About Widget 1.0 I think it still applies.)

When pressing cancel in view mode in the about widget, it will explicitly reset your changes, which I think is expected.
This Patch aligns both modes a bit more in terms of user experience.

Bug: T360062
Change-Id: Ic76b65d4f5c74a5f86e4b831897e711d1ac61928