ZObjectFactory::validatePersistentKeys: Disable use of JsonSchema, at least temporarily
This method (and not the later ones?) seems to be repeatedly triggering JsonSchema to
infinitely recurse on some view requests in production, though I can't reproduce the
breakage locally. Let's just do the simple key-existence-check on the stdClass, as it
is not going to explode.

It's possible that this will just kick the can down the road to the JsonSchema calls in
::create(), and in general we should just give up on validation, but this should at
least be a start.

Bug: T374241
Change-Id: I61318bce740398b3eae405b86465e52a24bee237