Do not recurse ZObject creation and validation inside Z99K1
Z99/Quote has one key that can contain any valid JSON value.
This means that when creating with ZObjectFactory::create, when
encountering a Z99K1, we should not recurse and create its content using
ZObjectFactory::create (which checks for validity), but assign directly.
For this, Z99K1 value should be considered terminal.

Bug: T380386
Change-Id: I1b12a6729cffa13e0dcc4dcbbeb8fdc072ae88d3