Use indexed page columns for join condition in preferred labels query
To order by creation date, the preferred labels query returns page_id
and joins on page_title=wlzl_zobject_zid and
page_content_model='zobject', so that we don't return talk pages.

This could be done with page_title=wlzl_zobject_zid and page_namespace=0
instead, as the result is the same, but page_namespace is indexed and
page_content_model is not.

Bug: T383561
Change-Id: If37e6b54f9f4ea6529d1eca13363bb5203b9403d