Add Wikidata Property components
Includes a new component for displaying Wikidata properties, and the
necessary store module to support it. This is part of the ongoing work
to support Wikidata properties in the Wikilambda UI.

Wikidata entity lookups now include a "load more" feature.
The ZObjectSelector has been updated to support this functionality as well.

Bug: T383643
Change-Id: I3ce06d4522f58367c8cb41bef1c7becb7d4741bb