ZObjectStore::findZTesterResult: If we get an invalid ZObject, delete it
This was a foreseen and handled error, but wasn't meant to happen often.
Now we're seeing it semi-frequently in production, let's delete each of
the mal-formed stored ZObjects as we encounter them.

Also, reduce the severity of this to a warning, as it's "self-healing"
unlike the other two branches, which remain errors as a human will have
to look into this and maybe fix it.

Change-Id: I152b9e5dfbfbc16a72d2822c5b3bf40fefd35287