Use Codex validation messages system supporting HTML on cdx-field now it exists
We might want to consider for the future to wrap all form-like fields in ZObjectSelector, ZObjectStringRenderer or ZCode in a cdx-field component.
Unsure what this will do to the UI and the zobject tree structure visually.
That would be a bigger refactor.

but that way we can get rid of these examples using custom fieldErrors with a cdx-message:
```
<cdx-message
	v-for="( error, index ) in fieldErrors"
	:key="'field-error-' + rowId + '-' + index"
	:type="error.type"
	:inline="true"
>
	<!-- eslint-disable vue/no-v-html -->
	<div v-html="getErrorMessage( error )"></div>
</cdx-message>
```

Bug: T388565
Change-Id: Ic5aa34945c3e7582bc6df042811ab65afdb5d12d