tests: Switch API tests from trying to call Beta Cluster to a local mock
This turns these disabled tests back on, at the cost of making them integration
tests, not end-to-end tests. They never now will actually hit an instance of
the function-orchestrator, just a mock of it; they let us confirm that, if the
orchestrator were to be called, the response would go through the API handler
as intended.

For actual end-to-end testing of this, we'll use the Catalyst system.

Bug: T360373
Bug: T374242
Bug: T371837
Change-Id: Ie61e8a0c8f501bd5f73d3c78609b9f73d7f26e9c