Stop pointing by default to the Beta Cluster, it's broken
At this point Beta Cluster is sufficiently broken that we should just stop
trying to point to it by default. This means the code is a smidgen less
useful out-of-the-box for testing, but in practice all our testing is done
locally or via scripts that set up the orchestrator anyway.

Bug: T374242
Change-Id: Ie61646b43ec546de500ea8a4c16505ac1635e1e6