WikiLambdaApiBase: Add logging for every remaining dieWith?(Z)Error
We were logging for most but not all of these, and irritatingly the
'response-malformed' one is the one we're hitting in production.

Also, fiddle with the priorities of the various log statements, as
the warning/error states are the ones that will come through in
production logging so focus on the issues we care about, and drop
logging of the 'validate' flag, always false.

Change-Id: I3b1afce07e313c55295c372ce196bc61401569b2
(cherry picked from commit a49e975144523a6a5e6dcd7bcafab905e767c846)