ZErrorTypeRegistry: Use the maybe-cached fetchZObject() instead of fetchZObjectByTitle()
NOTE: fetchZObject and fetchZObjectByTitle should not trigger more
fetches in case of exception, so they should not try throw new
ZErrorExceptions.
This is because ZError and ZErrorException request labels from all their
referred ZErrorTypes to build the messages, which need a lot more
fetches. Depending on the Zid that failed or the reason for the failure,
this can get into a nasty fetch infinite loop.
If we want fetchZObject and fetchZObjectByTitle to throw different
exceptions, we should create other internal system exceptions, and keep
ZErrorException for user responses.

Bug: T300530
Change-Id: I575f72b5be4bb83a425fd4b7bdafbb250a1cba56