Stop the perform_test api from passing errors to the validator function
When a test function call fails, we won't pass the error to the
validator. Instead, we will pass the response (Z24), similarly to when
the call is successful.
If we want to implement error equality for tests, we should do it
differently, and handle "expect exception" type of logic in the API,
instead of sending huge error objects generated by the orchestrator back
to the orchestrator.

Bug: T393156
Change-Id: I1263dbba3dcae8c8df6a66156a5e4a0fcc7dc60d