Don't trim arguments which are only whitespaces
getOrderedArgs can be called with an array of boolean flags, so that
different arguments may be trimmed or not depending on its particular
flag.
We can leverage this to trim all arguments except for those which are
only whitespaces, so "  foo " and "bar " will be trimmed but " " will
not.
This allows {{#function:JoinWithSeparator|hello|world| }} and similar
functions where a whitespace is needed.

Bug: T362251
Change-Id: I730cc328198bfd78dc61a7dcd892a6fba820515c