code: Don't use absolute class references in catches, and use Exception not Throwable
For consistency, we should declare every class we use at the top, and not have the
references be to absolute classes (in future, this will be enforced through the MW
coding standards).

Whilst we're at it, drop unused catch variable names (now we're in PHP 8.0+ only),
catch Exception rather than the more general and often not catch-appropriate class
Throwable (and fix logging to 'throwable' log entry which I think doesn't have the
special treatment in logstash).

Change-Id: I8e8275c9961d21da5beee68cf10cbea35ff466b8