WikifunctionsPFragmentSanitiserTokenHandler: Enable local links
User-generated HTML with <a> links will always be down-converted into
plain (escaped) text, unless the href is parsed and matches exactly to
the wgServer value or (on wikis with SiteMatrix installed, i.e.
Wikimedia production wikis) one of the other wikis on the cluster.

All attributes that aren't a 'local' link are dropped.

Note that links that are not allowed will also swallow the closing tag.

Bug: T398987
Change-Id: Id3f9b38bb17875865f346977649dab0a1012c5d6