Cache uses of Type and Function objects during FunctionCallHandler run
During execution of FunctionCallHandler we fetch the following objects
for validation of the input arguments and for building the function
call:
* Functions: We fetch the target function to inspect its arguments
* Types: We fetch types to get their read/display function Ids
* Other objects: When an object is referenced, we fetch it to validate
that its type matches the target input type.

We are now using fetchZObject, which checks the cache first. If there's
a miss, fetches the object from the DB and updates the object in the
cache.

Bug: T385619
Change-Id: Idf40b4090dd3a70f9b37d2ff1fd6025d47b041dd