Return terminal value of function call function only when reference
getZFunctionCallFunctionId should return terminal value of Z9 if Z7K1
has a reference. It should not return terminal values of nested Z7K1 or
Z18K1.

The only use case where nested values need to be checked is during
submission validation, and in order to do that, we need a different
validation function to return the failed fields and render an error.

Additionally, enable ArgumentReference component to render errors when
detected.

Bug: T398457
Change-Id: I734efdecf98ed53e6365e8aa0a598038aabdc6fb