Create job queue store module for scheduling parser/renderer calls
Parser and renderer calls (mostly renderer) can become extremely costly,
as an object that is loading in the UI can suddenly throw to the
orchestrator a big number of renderer calls at the same time.

This patch wraps renderer and parser calls in a simple job queue that
runs all the queued jobs one by one with a 200ms delay in between (this
should be adjusted after observing if the orchestrator handles this well
in production).

Bug: T397542
Change-Id: I4b7d629e7d27b18997d861b521ea075e7904ef3d