Fix FunctionInputParser flaky test
Validate events are used to disable and enable the dialog "done" button,
so for every action there are two "validate" events, one with a false on
start, and another one with false/true depending on the sucess or
failure of the validation.

Test needs to wait for the two first validate events, and overall wait
for four events, or else it will become flaky (e.g. depending on the
environment it might pass or fail)

This patch:
* makes the test more explicit
* set reactive data only when needed

Bug: T399845
Change-Id: I607d371d15a44b13f4996e807520d8475b6fedda